Support uuid pk.

+++ modules/boot-core/src/test/java/com/tairanchina/csp/dew/core/TxService.java
@@ -11,40 +11,40 @@ import org.springframework.transaction.annotation.Transactional;
 public class TxService {
 
     @Transactional
-    public void testCommit(){
+    public void testCommit() {
         BasicEntity basicEntity = new BasicEntity();
         basicEntity.setFieldA("TransactionA1");
         basicEntity.setFieldB("TransactionA1");
         Object id = Dew.ds().insert(basicEntity);
-        Assert.assertTrue((long)id != 0);
+        Assert.assertTrue((int) id != 0);
     }
 
-    @Transactional(rollbackFor = {Exception.class,RuntimeException.class,NullPointerException.class})
+    @Transactional(rollbackFor = {Exception.class, RuntimeException.class, NullPointerException.class})
     public void testRollBack() throws Exception {
         BasicEntity basicEntity = new BasicEntity();
         basicEntity.setFieldA("TransactionA2");
         basicEntity.setFieldB("TransactionA2");
         Object id = Dew.ds().insert(basicEntity);
-        Assert.assertTrue((long)id != 0);
+        Assert.assertTrue((int) id != 0);
         throw new Exception("");
     }
 
     @Transactional("test2TransactionManager")
-    public void testMultiCommit(){
+    public void testMultiCommit() {
         BasicEntity basicEntity = new BasicEntity();
         basicEntity.setFieldA("TransactionA1");
         basicEntity.setFieldB("TransactionA1");
         Object id = Dew.ds("test2").insert(basicEntity);
-        Assert.assertTrue((long)id != 0);
+        Assert.assertTrue((int) id != 0);
     }
 
-    @Transactional(value = ("test2TransactionManager"), rollbackFor = {Exception.class,RuntimeException.class,NullPointerException.class})
+    @Transactional(value = ("test2TransactionManager"), rollbackFor = {Exception.class, RuntimeException.class, NullPointerException.class})
     public void testMultiRollBack() throws Exception {
         BasicEntity basicEntity = new BasicEntity();
         basicEntity.setFieldA("TransactionA2");
         basicEntity.setFieldB("TransactionA2");
         Object id = Dew.ds("test2").insert(basicEntity);
-        Assert.assertTrue((long)id != 0);
+        Assert.assertTrue((int) id != 0);
         throw new Exception("");
     }
